+++ src/Orleans.CodeGeneration/CodeGeneratorCommon.cs
@@ -1,5 +1,3 @@
-using Microsoft.Extensions.Logging;
-
 namespace Orleans.CodeGenerator
 {
     using System;
